Refactor global navigation handling
