+++ src/components/global-navigation-handler/index.ts
@@ -1,0 +1,1 @@
+export { GlobalNavigationHandler } from './global-navigation-handler';
