+++ src/components/layout/layout.tsx
@@ -2,8 +2,7 @@ import { DesktopOnly } from '@components/desktop-only';
 import { DndOverlay } from '@components/dnd-overlay';
 import { useAddLocalFilesOrFolders } from '@hooks/use-add-local-files-folders';
 import { Stack } from '@mantine/core';
-import { useEffect } from 'react';
-import { Outlet, useNavigate } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 
 import { Header } from './components/header';
 
@@ -14,23 +13,6 @@ interface LayoutProps {
 
 export function Layout({ isFileAccessApiSupported, isMobileDevice }: LayoutProps) {
   const { handleFileDrop } = useAddLocalFilesOrFolders();
-  const navigate = useNavigate();
-
-  // Listen for navigation events from other parts of the app
-  useEffect(() => {
-    const handleNavigateEvent = (event: Event) => {
-      const customEvent = event as CustomEvent<{ route: string }>;
-      if (customEvent.detail?.route) {
-        navigate(customEvent.detail.route);
-      }
-    };
-
-    window.addEventListener('navigate-to-route', handleNavigateEvent);
-
-    return () => {
-      window.removeEventListener('navigate-to-route', handleNavigateEvent);
-    };
-  }, [navigate]);
 
   if (isMobileDevice) {
     return <DesktopOnly />;
